Added catch for large files

+++ server/src/Book/Book.routes.controller.ts
@@ -11,15 +11,20 @@ export class BookRoutesController {
     public static async addBook(req: Request, res: Response, next: NextFunction){
         if(Auth.isAuthenticated(req) && Auth.isAdmin(req)){
             const {author, title} = req.body;
-            let bookFilename = util.generateBookFilename({author, title})
+            let file = (<any>req).files.bookfile;
+            if(file.size <= 3000){
+                let bookFilename = util.generateBookFilename({author, title})
             
-            util.saveTxtAndMp3(bookFilename, req);
-
-            let savedBook = await BookRepository.save(new Book(author, title, req.session.user._id));
+                await util.saveTxtAndMp3(bookFilename, req);
     
-            res.sendStatus(200);
+                let savedBook = await BookRepository.save(new Book(author, title, req.session.user._id));
+        
+                res.sendStatus(200);
+            } else {
+                res.sendStatus(400);
+            }
         } else {
-            res.sendStatus(400)
+            res.sendStatus(401)
         }
     }
 
@@ -73,29 +78,33 @@ export class BookRoutesController {
 
             let newBookName;
             let files = (<any>req).files
-            const book: Book = await BookRepository.findOne(book_id);
-            let bookName = util.generateBookFilename({author: book.__author, title: book.__title});
-
-            if(author || title){
-                newBookName = util.generateBookFilename({
-                    author: (author ? author : book.__author),
-                    title: (title ? title : book.__title)
-                }); 
-
-                if(files){
-                    await util.saveTxtAndMp3(newBookName, req);
-                    await FileStorage.deleteFiles([bookName + ".txt", bookName + ".mp3"]);
+            if(files.bookfile.size <= 3000){
+                const book: Book = await BookRepository.findOne(book_id);
+                let bookName = util.generateBookFilename({author: book.__author, title: book.__title});
+    
+                if(author || title){
+                    newBookName = util.generateBookFilename({
+                        author: (author ? author : book.__author),
+                        title: (title ? title : book.__title)
+                    }); 
+    
+                    if(files){
+                        await util.saveTxtAndMp3(newBookName, req);
+                        await FileStorage.deleteFiles([bookName + ".txt", bookName + ".mp3"]);
+                    } else {
+                        await FileStorage.renameFile(bookName + ".txt", newBookName + ".txt");
+                        await FileStorage.renameFile(bookName + ".mp3", newBookName + ".mp3");
+                    }
+                    await BookRepository.updateOne(book_id, (author ? author : null), (title ? title : null));
                 } else {
-                    await FileStorage.renameFile(bookName + ".txt", newBookName + ".txt");
-                    await FileStorage.renameFile(bookName + ".mp3", newBookName + ".mp3");
+                    await util.saveTxtAndMp3(bookName, req);
                 }
-                await BookRepository.updateOne(book_id, (author ? author : null), (title ? title : null));
+                res.sendStatus(200);
             } else {
-                await util.saveTxtAndMp3(bookName, req);
+                res.sendStatus(400);
             }
-            res.sendStatus(200);
         } else {
-            res.sendStatus(400);
+            res.sendStatus(401);
         }
     }
